+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # ack-webpack - Change Log
 All notable changes to this project will be documented here.
 
+## [1.1.7] - 2017-03-15
+# Enhanced
+- screw IE8 for better compression
+# Fixed
+- modified webpack.config.js as production mode was failing
+
 ## [1.1.16] - 2017-03-09
 ### Added
 - new install process that targets installing jsDependency definitions of package.json file